[COMMS-925] Refactor so the TypeaheadFilter is not polluted
https://community.openproject.org/wp/COMMS-925

It's not really clear to me where this code should best live. I'd like
to have a Order class for it, much like the Queries::<Model>::Orders::<SpecificOrder> classes,
but that would require a major refactoring of the work package filters
which is out of scope at the moment.
