+++ spec/models/queries/work_packages/selects/exact_match_select_spec.rb
@@ -1,0 +1,125 @@
+# frozen_string_literal: true
+
+#-- copyright
+# OpenProject is an open source project management software.
+# Copyright (C) the OpenProject GmbH
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License version 3.
+#
+# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+# Copyright (C) 2006-2013 Jean-Philippe Lang
+# Copyright (C) 2010-2013 the ChiliProject Team
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
+# See COPYRIGHT and LICENSE files for more details.
+#++
+
+require "spec_helper"
+
+RSpec.describe Queries::WorkPackages::Selects::ExactMatchSelect do
+  describe ".condition_sql" do
+    subject(:sql) { described_class.condition_sql(query_string) }
+
+    def ranked_ids(ids)
+      WorkPackage.where(id: ids).order(Arel.sql("#{sql} DESC"), id: :asc).pluck(:id)
+    end
+
+    context "when the query string is blank" do
+      let(:query_string) { "" }
+
+      it { is_expected.to be_nil }
+    end
+
+    context "when the query string is whitespace-only" do
+      let(:query_string) { "   " }
+
+      it { is_expected.to be_nil }
+    end
+
+    context "when the query string has multiple words" do
+      let(:query_string) { "epic gorilla" }
+
+      it { is_expected.to be_nil }
+    end
+
+    context "when the query string is a plain numeric id" do
+      let!(:exact_work_package)  { create(:work_package) }
+      let!(:prefix_work_package) { create(:work_package) }
+      let(:query_string) { "5" }
+
+      before do
+        WorkPackage.where(id: exact_work_package.id).update_all(id: 5)
+        WorkPackage.where(id: prefix_work_package.id).update_all(id: 50)
+      end
+
+      it "ranks the exact numeric id above one that merely starts with it" do
+        expect(ranked_ids([5, 50]).first).to eq(5)
+      end
+    end
+
+    context "when the query string has a leading '#'" do
+      let!(:work_package) { create(:work_package) }
+      let(:query_string) { "##{work_package.id}" }
+
+      it "still matches the numeric id exactly" do
+        expect(sql).to include(work_package.id.to_s)
+      end
+    end
+
+    context "when the query string is a leading-zero numeric string" do
+      let(:query_string) { "007" }
+
+      it "does not raise" do
+        expect { sql }.not_to raise_error
+      end
+    end
+
+    context "when the query string is an exact semantic identifier" do
+      let!(:exact_work_package)  { create(:work_package) }
+      let!(:prefix_work_package) { create(:work_package) }
+      let!(:exact_alias) do
+        create(:work_package_semantic_alias, work_package: exact_work_package, identifier: "COM-5")
+      end
+      let!(:prefix_alias) do
+        create(:work_package_semantic_alias, work_package: prefix_work_package, identifier: "COM-50")
+      end
+      let(:query_string) { "COM-5" }
+
+      it "ranks the exact identifier match above one that only shares the prefix" do
+        expect(ranked_ids([exact_work_package.id, prefix_work_package.id]).first)
+          .to eq(exact_work_package.id)
+      end
+
+      context "when the query is given in lower case" do
+        let(:query_string) { "com-5" }
+
+        it "still matches COM-5" do
+          expect(ranked_ids([exact_work_package.id, prefix_work_package.id]).first)
+            .to eq(exact_work_package.id)
+        end
+      end
+
+      context "when matching a historical alias in classic mode",
+              with_settings: { work_packages_identifier: Setting::WorkPackageIdentifier::CLASSIC } do
+        it "still matches through the alias table" do
+          expect(ranked_ids([exact_work_package.id, prefix_work_package.id]).first)
+            .to eq(exact_work_package.id)
+        end
+      end
+    end
+  end
+end
